set up basic HTML text extraction
